+++ ZigBeeNet/ZCL/Field/ReadAttributeStatusRecordcs.cs
@@ -11,7 +11,7 @@ namespace ZigBeeNet.ZCL.Field
         /**
          * The attribute identifier.
          */
-        public int AttributeIdentifier { get; private set; }
+        public ushort AttributeIdentifier { get; private set; }
 
         /**
          * The status.
@@ -39,7 +39,7 @@ namespace ZigBeeNet.ZCL.Field
 
         public void Deserialize(IZigBeeDeserializer deserializer)
         {
-            AttributeIdentifier = (int)deserializer.ReadZigBeeType(ZclDataType.Get(DataType.UNSIGNED_16_BIT_INTEGER));
+            AttributeIdentifier = (ushort)deserializer.ReadZigBeeType(ZclDataType.Get(DataType.UNSIGNED_16_BIT_INTEGER));
             Status = (ZclStatus)deserializer.ReadZigBeeType(ZclDataType.Get(DataType.ZCL_STATUS));
             if (Status == ZclStatus.SUCCESS)
             {
